feat: stop relays deregistering on close
The conductor service already monitors heartbeats and reassigns projects
to new relays if one dies. De-registering on close means that the relay
would have to be in control of re-assigning it's own projects.

The delay of 15-30 seconds where a relay is unreachable should be fine,
especially as the request will still be stored and can be pushed to a
new relay server when available